move sidecar trait to new file
